Various updates: 1) updated Ontology to latest version (working draft) 2) fixed getEvent and setEvent to get/setEvents (plural) in LogisticsObjectsServiceImpl
We had a persistent error due to dependencies between the API and Cargo ontologies. JOPA created models for all API classes INCLUDING those that are referenced to the cargo ontology. This created a duplicate for Company in both API and Cargo models. The solution was to remove the (fake) Company class in the API model and correct references to Company in several files.
A related issue was that these duplicate/fake classes for Company, Branch and LogisticsObject included duplicate "toString" methods that caused an error. These were removed.

Possibly these errors will be replicated when we do a new update of the ontology. We are not sure yet how to link these ontologies correctly so we'll just have to repeat the above if necessary.